scene: Only stop LEDs for running scenes.
When multiple scenes share LEDs, and one is running, the other scenes
should not cancel the timers for those LEDs when being paused or stopped.

Cancelling those timers (calling the stop_scene_leds function) will break
any blink or fade operations on the running scene.

Overall setup for examples:
fast_blink, slow_blink, and solid are three scenes on the same LED.
fast_blink is the highest priority scene, slow_blink is medium priority,
and solid is the lowest priority scene.

Example 1: fast_blink is running, slow_blink is paused by fast_blink, and
solid is paused by slow_blink.
The condition which triggers slow_blink has ended, so the scene is stopped
via the ubus method.  The fast_blink operation should not be interrupted
in this case, either by slow_blink being stopped, or solid being paused again.

Example 2: fast_blink is running, and slow_blink is idle.
The condition which triggers slow_blink has started, so the scene is
started via the ubus method.  Since slow_blink is of lower priority, it
is immediately paused.  The fast_blink operation should not be interrupted
in this case.

Signed-off-by: Joe Stefek <jstefek@datto.com>