scene: Update all LEDs attached to a scene from the same timer handler
On lower powered CPUs, when using the "blink" option for a scene with
multiple LEDs, the different timers for each LED will slowly get out of sync.
On an ipq40xx sitting idle, this can happen as quickly as after a couple
minutes.  To address this, when LEDs are added as a part of a scene, use a
timer for the scene as a whole, rather than for each individual LED.

Example use case: A trio of red/green/blue LEDs comes through the same light
pipe, and is used for various statuses.  When trying to blink these three
LEDs together, to provide a blinking white, the different colors will drift
out of sync and provide noticeable points before and after the "white" period
where individual colors are lit.